Rework some stuff so the bot can run in the future  * Get rid of custom Telegram.Bot  * Get rid of Telegram.Bot.Core  * Install Telegram.Bot v16 everywhere (v17 not compatible)  * Get rid of a lot of build errors caused by all this

+++ Werewolf for Telegram/Database/GlobalBan.cs
@@ -15,7 +15,7 @@ namespace Database
     public partial class GlobalBan
     {
         public int Id { get; set; }
-        public int TelegramId { get; set; }
+        public long TelegramId { get; set; }
         public string Reason { get; set; }
         public System.DateTime Expires { get; set; }
         public string BannedBy { get; set; }
