+++ Werewolf for Telegram/Werewolf Node/Models/IPlayer.cs
@@ -28,9 +28,9 @@ namespace Werewolf_Node.Models
         /// <summary>
         /// Choice of the player they want to use their ability on
         /// </summary>
-        public int Choice { get; set; } = 0;
+        public long Choice { get; set; } = 0;
 
-        public int Choice2 { get; set; } = 0;
+        public long Choice2 { get; set; } = 0;
 
         /// <summary>
         /// Whooops! you died...
@@ -78,7 +78,7 @@ namespace Werewolf_Node.Models
         //public bool HasNightAction { get; set; } = false;
         //public bool HasDayAction { get; set; } = false;
         public int DayCult { get; set; } = 0;
-        public int RoleModel { get; set; } = 0;
+        public long RoleModel { get; set; } = 0;
         [JsonConverter(typeof(StringEnumConverter))]
         public IRole KilledByRole { get; set; }
         public bool DiedByVisitingKiller { get; set; } = false;
@@ -92,7 +92,7 @@ namespace Werewolf_Node.Models
         public int MessageId { get; set; }
         public string Name { get; set; }
         public bool InLove { get; set; } = false;
-        public int LoverId { get; set; } = 0;
+        public long LoverId { get; set; } = 0;
         public bool Doused { get; set; } = false;
         public bool Burning { get; set; } = false;
         public int DBPlayerId { get; set; } = 0;
@@ -106,7 +106,7 @@ namespace Werewolf_Node.Models
         public string Language { get; set; } = "English";
         public bool Won { get; set; } = false;
 
-        public int Id;
+        public long Id;
         [JsonConverter(typeof(StringEnumConverter))]
         public QuestionAsked CurrentQuestion { get; set; }
 
@@ -128,7 +128,7 @@ namespace Werewolf_Node.Models
         public bool LastShotWasSKWolf { get; set; } = false;
         public bool DoubleKillEnding { get; set; } = false;
         public bool Bitten { get; set; }
-        public List<int> CorrectSnooped { get; set; } = new List<int>();
+        public List<long> CorrectSnooped { get; set; } = new List<long>();
         public bool SpeedDating { get; set; } = false;
         public int FirstStone { get; set; } = 0;
         public int CHHuntedCultCount { get; set; } = 0;
